Updated Javadocs and added TTS support
Fixed
- Irrelevant javadocs since it's inherited. Copilot committed this terrible mistake. I need somebody to blame ¯\_(ツ)_/¯ (kidding potential readers)

Added
- TTS support

I wasted like 88% of the time trying to debug an issue as to why my TTS was in a deeper voice. Turns out, the sample rate is to fault. I set it to 16000 when I was testing STT.

First I thought the model somehow corruptegd (Doesn't really make sense, but I was desperate). Then I thought it may be my USB speaker (then my computer speaker), then my headphones. Then I thought it was Linux driver (But realized it wouldn't be because VLC plays it normally). Then I thought it was IDEA. Then atlast, I **accidentally** clicked on a file that displayed the sample rate of the model (22050), and I cried really hard).